move remaining stores to apps #377
