+++ app/js/store/reducers/index.js
@@ -1,9 +1,9 @@
 import { combineReducers } from 'redux'
 
-import { AccountReducer }  from '../account'
-import { AuthReducer }     from '../auth'
+import { AccountReducer }  from '../../account/store/account'
+import { AuthReducer }     from '../../auth/store/auth'
 import { SettingsReducer } from '../../account/store/settings'
-import { IdentityReducer } from '../identities'
+import { IdentityReducer } from '../../profiles/store/identities'
 import { SearchReducer }   from '../../profiles/store/search'
 
 const RootReducer = combineReducers({
